frontend-app/requirements.txt
streamlit
google-cloud-bigquery
pandas
db-dtypes
